% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\docType{methods}
\name{apply,LinkedMatrix-method}
\alias{apply,LinkedMatrix-method}
\title{Apply function for \code{\linkS4class{ColumnLinkedMatrix}} or
\code{\linkS4class{RowLinkedMatrix}} objects.}
\usage{
\S4method{apply}{LinkedMatrix}(X, MARGIN, FUN, chunkSize = 1000,
  verbose = FALSE, ...)
}
\arguments{
\item{X}{Either a \code{\linkS4class{ColumnLinkedMatrix}} or a
\code{\linkS4class{RowLinkedMatrix}} object.}

\item{MARGIN}{Use 1 to obtain row summaries or 2 to obtain column summaries.}

\item{chunkSize}{The number of columns or rows that are processed at a time
(see Details).}
}
\value{
Returns a \code{matrix} or a \code{list} with results from FUN.
}
\description{
This function brings chunks of data (of size \code{chunkSize}) from the
distributed list into RAM as \code{matrix} objects and calls \code{apply} of
the base package to obtain the summaries for the chunk. Results from all the
chunks are collected and returned.
}

