% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColumnLinkedMatrix.R
\name{as.ColumnLinkedMatrix}
\alias{as.ColumnLinkedMatrix}
\title{Converts an Object to a ColumnLinkedMatrix Object.}
\usage{
as.ColumnLinkedMatrix(x, ...)
}
\arguments{
\item{x}{An object to convert to a \linkS4class{ColumnLinkedMatrix} object.}

\item{...}{Additional arguments.}
}
\value{
A \linkS4class{ColumnLinkedMatrix} object.
}
\description{
Converts an Object to a ColumnLinkedMatrix Object.
}
