% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowLinkedMatrix.R
\name{as.RowLinkedMatrix}
\alias{as.RowLinkedMatrix}
\title{Converts an Object to a RowLinkedMatrix Object.}
\usage{
as.RowLinkedMatrix(x, ...)
}
\arguments{
\item{x}{An object to convert to a \linkS4class{RowLinkedMatrix} object.}

\item{...}{Additional arguments.}
}
\value{
A \linkS4class{RowLinkedMatrix} object.
}
\description{
Converts an Object to a RowLinkedMatrix Object.
}
