% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{chunks}
\alias{chunks}
\title{Provides information about how data is distributed into binary files.}
\usage{
chunks(x)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ColumnLinkedMatrix}} or a
\code{\linkS4class{RowLinkedMatrix}} object}
}
\value{
A matrix with information per chunk in rows.
}
\description{
Column-distributed (\code{\linkS4class{ColumnLinkedMatrix}}) and row-distributed
matrices (\code{\linkS4class{RowLinkedMatrix}}) have the content of the array mapped
to possibly multiple binary files. Each chunk is an \code{ff_matrix} object.
\code{chunks} gives, for each chunk, the row or column indexes at which each
chunk start and ends.
}

