% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\docType{class}
\name{LinkedMatrix-class}
\alias{LinkedMatrix-class}
\title{A Class Union of ColumnLinkedMatrix and RowLinkedMatrix.}
\description{
This class is abstract and no objects can be created from it. It can be used
to check whether an object is either of type \linkS4class{ColumnLinkedMatrix} or
of type \linkS4class{RowLinkedMatrix} using \code{is(x, "LinkedMatrix")} and to assign
methods for both \code{ColumnLinkedMatrix} and \code{RowLinkedMatrix} classes, e.g.
\code{show}.
}
\section{Methods}{

\itemize{
\item \code{length}
\item \code{as.matrix}
\item \code{show}
}
}

\seealso{
\linkS4class{ColumnLinkedMatrix} and \linkS4class{RowLinkedMatrix} for
implementations of column-linked and row-linked matrices, respectively.
}
