% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{nNodes}
\alias{nNodes}
\title{Returns the Number of Nodes.}
\usage{
nNodes(x)
}
\arguments{
\item{x}{Either a \linkS4class{ColumnLinkedMatrix} or a \linkS4class{RowLinkedMatrix}
object.}
}
\value{
The number of nodes.
}
\description{
Returns the Number of Nodes.
}
