\name{nodes}
\alias{nodes}
\title{Returns the Column or Row Indexes at Which Each Node Starts and Ends}
\description{
Returns the Column or Row Indexes at Which Each Node Starts and Ends.
}
\usage{
nodes(x)
}
\arguments{
  \item{x}{
    Either a \code{ColumnLinkedMatrix} or a \code{RowLinkedMatrix} object.
  }
}
\value{
A matrix.
}
\seealso{
\code{\link{ColumnLinkedMatrix-class}}, \code{\link{RowLinkedMatrix-class}},
and \code{\link{LinkedMatrix-class}} for more information on the
\code{ColumnLinkedMatrix}, \code{RowLinkedMatrix}, and \code{LinkedMatrix}
classes.
}
