% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowLinkedMatrix.R
\name{rbind.RowLinkedMatrix}
\alias{rbind.RowLinkedMatrix}
\title{Combine Matrix-Like Objects by Rows.}
\usage{
\method{rbind}{RowLinkedMatrix}(..., deparse.level = 1)
}
\arguments{
\item{...}{Matrix-like objects to be combined by rows.}

\item{deparse.level}{Currently unused, defaults to 0.}
}
\description{
Compared to the \code{[RowLinkedMatrix][initialize,RowLinkedMatrix-method]}
constructor, nested \linkS4class{LinkedMatrix} objects that are passed via \code{...}
will not be treated as matrix-like objects, but their nodes will be
extracted and merged with the new \linkS4class{RowLinkedMatrix} object for a more
compact representation.
}
