% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RowLinkedMatrix.R
\docType{class}
\name{RowLinkedMatrix-class}
\alias{RowLinkedMatrix}
\alias{RowLinkedMatrix-class}
\title{An S4 class to represent a row-distributed \code{mmMatrix}}
\description{
\code{RowLinkedMatrix} inherits from \code{\link{list}}. Each element of the list is
an \code{ff_matrix} object.
}

