\docType{methods}
\name{show,LinkedMatrix-method}
\alias{show,LinkedMatrix-method}
\title{Show a LinkedMatrix Object}
\description{
Display the object, by printing, plotting or whatever suits its class.
}
\usage{
\S4method{show}{LinkedMatrix}(object)
}
\arguments{
  \item{object}{
    A \code{LinkedMatrix} object.
  }
}
\seealso{
\code{\link{ColumnLinkedMatrix-class}}, \code{\link{RowLinkedMatrix-class}},
and \code{\link{LinkedMatrix}} for more information on the
\code{ColumnLinkedMatrix}, \code{RowLinkedMatrix}, and \code{LinkedMatrix}
classes.
}
