% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{as.matrix.LinkedMatrix}
\alias{as.matrix.LinkedMatrix}
\title{Converts a \code{\link[=LinkedMatrix-class]{LinkedMatrix}} instance to a
\code{matrix} (if small enough).}
\usage{
\method{as.matrix}{LinkedMatrix}(x, ...)
}
\arguments{
\item{x}{Either a \code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}}
or a \code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Converts a \code{\link[=LinkedMatrix-class]{LinkedMatrix}} instance to a
\code{matrix} (if small enough).
}

