% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColumnLinkedMatrix.R
\name{cbind.ColumnLinkedMatrix}
\alias{cbind.ColumnLinkedMatrix}
\title{Combine matrix-like objects by columns.}
\usage{
cbind.ColumnLinkedMatrix(..., deparse.level = 0)
}
\arguments{
\item{...}{Matrix-like objects to be combined by columns.}

\item{deparse.level}{Currently unused, defaults to 0.}
}
\description{
Compared to the
\code{\link[=initialize,ColumnLinkedMatrix-method]{ColumnLinkedMatrix}}
constructor, nested \code{\link[=LinkedMatrix-class]{LinkedMatrix}} objects
that are passed via \code{...} will not be treated as matrix-like objects,
but their nodes will be extracted and merged with the new
\code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}} object for a
more compact representation.
}

