% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowLinkedMatrix.R
\name{cbind.RowLinkedMatrix}
\alias{cbind.RowLinkedMatrix}
\title{Combine matrix-like objects by columns.}
\usage{
cbind.RowLinkedMatrix(..., deparse.level = 0)
}
\arguments{
\item{...}{Matrix-like objects to be combined by columns.}

\item{deparse.level}{Currently unused, defaults to 0.}
}
\description{
This method is currently undefined for
\code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} objects.
}

