% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{chunks}
\alias{chunks}
\title{Provides information about how data is distributed.}
\usage{
chunks(x)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ColumnLinkedMatrix}} or a
\code{\linkS4class{RowLinkedMatrix}} object}
}
\value{
A matrix with information per chunk in rows.
}
\description{
\code{chunks} gives, for each chunk, the row or column indexes at which each
chunk starts and ends.
}

