% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColumnLinkedMatrix.R
\docType{class}
\name{ColumnLinkedMatrix-class}
\alias{ColumnLinkedMatrix-class}
\alias{ColumnLinkedMatrix}
\title{An S4 Class to Represent a Column-Linked LinkedMatrix.}
\description{
This class treats a list of matrix-like objects that are linked together by
columns and have the same number of rows similarly to a regular \code{matrix} by
implementing key methods such as \code{[} and \code{[<-} for extracting and replacing
matrix elements, \code{dim} to retrieve dimensions, and \code{dimnames} and
\code{dimnames<-} to retrieve and set dimnames. Each list element is called a
node and can be extracted or replaced using \code{[[} and \code{[[<-}. A matrix-like
object is one that has two dimensions and implements at least \code{dim} and \code{[}.
}
\details{
There are several ways to create an instance of this class: either by using
one of the constructors
\code{[ColumnLinkedMatrix(...)][initialize,ColumnLinkedMatrix-method]} or
\code{[new("ColumnLinkedMatrix", ...)][initialize,ColumnLinkedMatrix-method]}, or
by using the more general \code{\link[=LinkedMatrix]{LinkedMatrix()}} function that constructs objects
of certain dimensions with a configurable number and type of nodes.
}
