% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ColumnLinkedMatrix.R
\docType{class}
\name{ColumnLinkedMatrix-class}
\alias{ColumnLinkedMatrix}
\alias{ColumnLinkedMatrix-class}
\title{An S4 class to represent a column-distributed \code{mmMatrix}.}
\description{
\code{ColumnLinkedMatrix} inherits from \code{\link{list}}. Each element of the list is
an \code{ff_matrix} object.
}

