% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ColumnLinkedMatrix.R
\docType{class}
\name{ColumnLinkedMatrix-class}
\alias{ColumnLinkedMatrix}
\alias{ColumnLinkedMatrix-class}
\title{An S4 class to represent a column-linked \code{LinkedMatrix}.}
\description{
\code{ColumnLinkedMatrix} inherits from \code{\link{list}}.
}

