% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{index}
\alias{index}
\title{Finds the position of a set of rows or columns in a
\code{\linkS4class{ColumnLinkedMatrix}} or a
\code{\linkS4class{RowLinkedMatrix}} object.}
\usage{
index(x)
}
\arguments{
\item{x}{Either a \code{\linkS4class{ColumnLinkedMatrix}} or a
\code{\linkS4class{RowLinkedMatrix}} object}
}
\value{
A matrix with information in which chunks each row and column of the
  matrix is located.
}
\description{
Finds the position of a set of rows or columns in a
\code{\linkS4class{ColumnLinkedMatrix}} or a
\code{\linkS4class{RowLinkedMatrix}} object.
}

