% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{index}
\alias{index}
\title{Maps each column or row index of a linked matrix to the column or row index
of its corresponding node.}
\usage{
index(x, ...)
}
\arguments{
\item{x}{Either a \linkS4class{ColumnLinkedMatrix} or a \linkS4class{RowLinkedMatrix}
object.}

\item{...}{Additional arguments (see Details).}
}
\value{
A matrix.
}
\description{
If \code{j} for \linkS4class{ColumnLinkedMatrix} or \code{i} for \linkS4class{RowLinkedMatrix} is
passed, it will only generate entries for the given indices.
}
