% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{index}
\alias{index}
\title{Maps each column or row index of a linked matrix to the column or row index
of its corresponding node.}
\usage{
index(x)
}
\arguments{
\item{x}{Either a \code{\linkS4class{ColumnLinkedMatrix}} or a
\code{\linkS4class{RowLinkedMatrix}} object}
}
\value{
A matrix.
}
\description{
Maps each column or row index of a linked matrix to the column or row index
of its corresponding node.
}

