% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{index}
\alias{index}
\title{Maps Each Column or Row Index of a Linked Matrix to the Column or Row Index
of Its Corresponding Node.}
\usage{
index(x, ...)
}
\arguments{
\item{x}{Either a \linkS4class{ColumnLinkedMatrix} or a \linkS4class{RowLinkedMatrix}
object.}

\item{...}{Additional arguments (see Details).}
}
\value{
A matrix.
}
\description{
If \code{j} for \linkS4class{ColumnLinkedMatrix} or \code{i} for \linkS4class{RowLinkedMatrix} is
passed, it will only generate entries for the given indices. \code{sort}, which
is set by default, determines whether \code{j} or \code{i} should be sorted before
building the index.
}
