% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColumnLinkedMatrix.R
\docType{methods}
\name{initialize,ColumnLinkedMatrix-method}
\alias{initialize,ColumnLinkedMatrix-method}
\title{Creates a new \code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}}
instance.}
\usage{
\S4method{initialize}{ColumnLinkedMatrix}(.Object, ...)
}
\arguments{
\item{.Object}{The
\code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}} instance to be
initialized.}

\item{...}{objects, possibly named.}
}
\description{
Creates a new \code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}}
instance.
}

