% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowLinkedMatrix.R
\docType{methods}
\name{initialize,RowLinkedMatrix-method}
\alias{initialize,RowLinkedMatrix-method}
\title{Creates a new \code{\linkS4class{RowLinkedMatrix}} instance.}
\usage{
\S4method{initialize}{RowLinkedMatrix}(.Object, ...)
}
\arguments{
\item{.Object}{The \code{\linkS4class{RowLinkedMatrix}} instance to be initialized.}

\item{...}{objects, possibly named.}
}
\description{
Creates a new \code{\linkS4class{RowLinkedMatrix}} instance.
}

