% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowLinkedMatrix.R
\docType{methods}
\name{initialize,RowLinkedMatrix-method}
\alias{initialize,RowLinkedMatrix-method}
\title{Creates a new \code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} instance.}
\usage{
\S4method{initialize}{RowLinkedMatrix}(.Object, ...)
}
\arguments{
\item{.Object}{The \code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}}
instance to be initialized. This argument is passed in by R and can be
ignored, but still needs to be documented.}

\item{...}{A sequence of matrix-like objects of the same column-dimension.}
}
\description{
This method is run when a
\code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} object is created using
\code{\link[=initialize,RowLinkedMatrix-method]{RowLinkedMatrix(...)}} or
\code{\link[=initialize,RowLinkedMatrix-method]{new("RowLinkedMatrix",...)}}
and accepts a list of matrix-like objects as \code{...}. A matrix-like object
is one that has two dimensions and implements at least \code{dim} and
\code{[}. Each object needs to have the same number of columns to be linked
together. \code{\link[=LinkedMatrix-class]{LinkedMatrix}} can be nested as
long as they are conformable. If no matrix-like objects are given, a single
1x1 node of type \code{matrix} filled with \code{NA} is returned.
}

