% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RowLinkedMatrix.R
\docType{methods}
\name{initialize,RowLinkedMatrix-method}
\alias{initialize,RowLinkedMatrix-method}
\title{Creates a new RowLinkedMatrix instance.}
\usage{
\S4method{initialize}{RowLinkedMatrix}(.Object, ...)
}
\arguments{
\item{.Object}{The \code{RowLinkedMatrix} instance to be initialized.}

\item{...}{objects, possibly named.}
}
\description{
Creates a new RowLinkedMatrix instance.
}

