% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\docType{class}
\name{LinkedMatrix-class}
\alias{LinkedMatrix-class}
\title{An abstract S4 class to represent linked matrices.}
\description{
An abstract S4 class to represent linked matrices.
}
\seealso{
\code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}} or 
  \code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} for implementations
  of column-linked matrices or row-linked matrices, respectively.
}

