% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\docType{class}
\name{LinkedMatrix-class}
\alias{LinkedMatrix-class}
\title{An abstract S4 class union of
\code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}} and
\code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}}.}
\description{
This class is a class union and can therefore not be initialized. It can be
used to check whether an object is either of type
\code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}} or of type
\code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} using \code{is(x,
"LinkedMatrix")}, and to assign methods for both
\code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}} and
\code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} classes, e.g.
\code{\link[=show,LinkedMatrix-method]{show}}.
}
\seealso{
\code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}} or
  \code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} for implementations
  of column-linked matrices or row-linked matrices, respectively.
}

