% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\docType{class}
\name{LinkedMatrix-class}
\alias{LinkedMatrix-class}
\title{An abstract S4 class to represent linked matrices.}
\description{
An abstract S4 class to represent linked matrices.
}
\seealso{
\code{\linkS4class{ColumnLinkedMatrix}} or
  \code{\linkS4class{RowLinkedMatrix}} for implementations of column-linked
  matrices or row-linked matrices, respectively.
}

