\docType{package}
\name{LinkedMatrix-package}
\alias{LinkedMatrix-package}
\title{\packageTitle{LinkedMatrix}}
\description{\packageDescription{LinkedMatrix}}
\seealso{
    \code{\link{ColumnLinkedMatrix-class}},
    \code{\link{RowLinkedMatrix-class}}, and \code{\link{LinkedMatrix-class}}
    for more information on the \code{ColumnLinkedMatrix},
    \code{RowLinkedMatrix}, and \code{LinkedMatrix} classes.
}
