% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{LinkedMatrix}
\alias{LinkedMatrix}
\alias{LinkedMatrix-package}
\title{LinkedMatrix: Column-Linked and Row-Linked Matrices}
\description{
Matrices implemented as collections of matrix-like nodes, linked by columns or rows.
}

