% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{nNodes}
\alias{nNodes}
\title{Returns the number of nodes.}
\usage{
nNodes(x)
}
\arguments{
\item{x}{Either a \code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}}
or a \code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} object.}
}
\description{
Returns the number of nodes.
}

