\name{nodes}
\alias{nodes}
\title{Returns the Column or Row Indexes at Which Each Node Starts and Ends}
\usage{
nodes(x)
}
\arguments{
\item{x}{Either a \linkS4class{ColumnLinkedMatrix} or a \linkS4class{RowLinkedMatrix}
object.}
}
\value{
A matrix.
}
\description{
Returns the Column or Row Indexes at Which Each Node Starts and Ends.
}
