% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{nodes}
\alias{nodes}
\title{Returns the column or row indexes at which each node starts and ends.}
\usage{
nodes(x)
}
\arguments{
\item{x}{Either a \code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}}
or a \code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} object.}
}
\value{
A matrix.
}
\description{
Returns the column or row indexes at which each node starts and ends.
}

