% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\name{nodes}
\alias{nodes}
\title{Returns the column or row indexes at which each node starts and ends.}
\usage{
nodes(x)
}
\arguments{
\item{x}{Either a \code{\linkS4class{ColumnLinkedMatrix}} or a
\code{\linkS4class{RowLinkedMatrix}} object}
}
\value{
A matrix.
}
\description{
Returns the column or row indexes at which each node starts and ends.
}

