% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColumnLinkedMatrix.R
\name{rbind.ColumnLinkedMatrix}
\alias{rbind.ColumnLinkedMatrix}
\title{Combine matrix-like objects by rows.}
\usage{
rbind.ColumnLinkedMatrix(..., deparse.level = 1)
}
\arguments{
\item{...}{Matrix-like objects to be combined by rows.}

\item{deparse.level}{Currently unused, defaults to 0.}
}
\description{
Combine matrix-like objects by rows.
}

