% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowLinkedMatrix.R
\name{rbind.RowLinkedMatrix}
\alias{rbind.RowLinkedMatrix}
\title{Combine matrix-like objects by rows.}
\usage{
rbind.RowLinkedMatrix(..., deparse.level = 1)
}
\arguments{
\item{...}{Matrix-like objects to be combined by rows.}

\item{deparse.level}{Currently unused, defaults to 0.}
}
\description{
Compared to the
\code{\link[=initialize,RowLinkedMatrix-method]{RowLinkedMatrix}}
constructor, nested \code{\link[=LinkedMatrix-class]{LinkedMatrix}} objects
that are passed via \code{...} will not be treated as matrix-like objects,
but their nodes will be extracted and merged with the new
\code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} object for a more
compact representation.
}

