% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowLinkedMatrix.R
\docType{class}
\name{RowLinkedMatrix-class}
\alias{RowLinkedMatrix}
\alias{RowLinkedMatrix-class}
\title{An S4 class to represent a row-linked
\code{\link[=LinkedMatrix-class]{LinkedMatrix}}.}
\description{
\code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} inherits from
\code{\link{list}}.
}

