% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\docType{methods}
\name{show,LinkedMatrix-method}
\alias{show,LinkedMatrix-method}
\title{Show a \linkS4class{LinkedMatrix} object.}
\usage{
\S4method{show}{LinkedMatrix}(object)
}
\arguments{
\item{object}{Either a \linkS4class{ColumnLinkedMatrix} or a
\linkS4class{RowLinkedMatrix} object.}
}
\description{
This method is run when a \linkS4class{LinkedMatrix} object is printed.
}
