% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\docType{methods}
\name{show,LinkedMatrix-method}
\alias{show,LinkedMatrix-method}
\title{Show a \code{\link[=LinkedMatrix-class]{LinkedMatrix}} object.}
\usage{
\S4method{show}{LinkedMatrix}(object)
}
\arguments{
\item{object}{Either a
\code{\link[=ColumnLinkedMatrix-class]{ColumnLinkedMatrix}} or a
\code{\link[=RowLinkedMatrix-class]{RowLinkedMatrix}} object.}
}
\description{
This method is run when a \code{\link[=LinkedMatrix-class]{LinkedMatrix}}
object is printed.
}

