% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\docType{methods}
\name{show,LinkedMatrix-method}
\alias{show,LinkedMatrix-method}
\title{Show a LinkedMatrix Object.}
\usage{
\S4method{show}{LinkedMatrix}(object)
}
\arguments{
\item{object}{A \linkS4class{LinkedMatrix} object.}
}
\description{
Display the object, by printing, plotting or whatever suits its class.
}
