% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinkedMatrix.R
\docType{methods}
\name{show,LinkedMatrix-method}
\alias{show,LinkedMatrix-method}
\title{Show a LinkedMatrix object.}
\usage{
\S4method{show}{LinkedMatrix}(object)
}
\arguments{
\item{object}{Either a \code{\linkS4class{ColumnLinkedMatrix}} or a
\code{\linkS4class{RowLinkedMatrix}} object.}
}
\description{
Show a LinkedMatrix object.
}

